Merged v0.6 to master

+++ src/util/RaceTime.h
@@ -1,7 +1,7 @@
 
 #pragma once
 
-long RaceTime_string_to_ms(char* time_string);
-char* RaceTime_ms_to_string(long time);
+unsigned int RaceTime_string_to_ms(char* time_string);
+char* RaceTime_ms_to_string(unsigned int time);
 char* RaceTime_percentage_to_string(float value);
